// frontend/src/backend/api/backend.ts
export enum Kind {
    collective = "collective",
    personnal = "personnal"
}
export enum Priority {
    low = "low",
    high = "high",
    medium = "medium"
}
export enum Status {
    done = "done",
    todo = "todo",
    inProgress = "inProgress"
}